strawberry_detection_final/requirements.txt
blinker==1.8.2
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1687884041762/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1723018708242/work
charset-normalizer==3.3.2
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
efficientnet_pytorch==0.7.1
filelock==3.16.1
Flask==3.0.3
fonttools==4.57.0
fsspec==2025.3.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hf-xet==1.1.3
hpack==4.0.0
huggingface-hub==0.32.4
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1726459485162/work
imagecodecs @ file:///home/conda/feedstock_root/build_artifacts/imagecodecs_1679699601060/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1729190692267/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader @ file:///home/conda/feedstock_root/build_artifacts/lazy-loader_1723774329602/work
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
munch==4.0.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1680692919326/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1687808334761/work
opencv-python==4.11.0.86
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.0.3
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1688256056872/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1717777836653/work
pretrainedmodels==0.7.4
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pyparsing==3.1.4
pyrealsense2==2.55.1.6486
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets @ file:///home/conda/feedstock_root/build_artifacts/pywavelets_1695567581506/work
PyYAML==6.0.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1717057054362/work
safetensors==0.5.3
scikit-image @ file:///home/conda/feedstock_root/build_artifacts/scikit-image_1688804519263/work/dist/scikit_image-0.21.0-cp38-cp38-linux_aarch64.whl#sha256=a8e9ed0149b727e0814068be5d4939042e4508d7d9a1b444ce4b972f7366d757
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1683900222033/work/base/dist/scipy-1.10.1-cp38-cp38-linux_aarch64.whl#sha256=79d95615707b668c8962067d4b92679982d8071638a200049f9c69c916da59c6
seaborn==0.13.2
segmentation-models-pytorch==0.3.3
six==1.17.0
sympy==1.13.3
tifffile @ file:///home/conda/feedstock_root/build_artifacts/tifffile_1689063654925/work
timm==0.9.2
torch @ file:///home/dfx/torch-2.1.0a0%2B41361538.nv23.06-cp38-cp38-linux_aarch64.whl#sha256=5112e2ef5051f1003ae2ffb545ae596377df66979d62a954d774f18009064dbe
torchvision==0.16.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.149
ultralytics-thop==2.0.14
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1726496430923/work
Werkzeug==3.0.6
zipp==3.20.2
zstandard @ file:///home/conda/feedstock_root/build_artifacts/zstandard_1667296135032/work
